+++ packages/tinacms/src/admin/components/AuthTemplate.tsx
@@ -24,7 +24,7 @@ const AuthTemplate = ({
     <div className="h-screen w-full bg-gradient-to-b from-blue-900 to-gray-900 flex items-center justify-center px-4 py-6">
       <div className="bg-white rounded-lg overflow-hidden shadow-lg w-full max-w-lg">
         <div className="px-5 py-4 border-b border-gray-150">
-          <h2 className="text-2xl tracking-wide text-gray-700 flex items-center gap-0.5">
+          <h2 className="text-2xl font-sans tracking-wide text-gray-700 flex items-center gap-0.5">
             <svg
               viewBox="0 0 32 32"
               fill="#EC4815"