+++ packages/@tinacms/toolkit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.56.31
+
+### Patch Changes
+
+- 999f0895a: Set font family on heading elements
+
+## 0.56.30
+
+### Patch Changes
+
+- aaaa5bb09: Added pagination to the CMS
+
 ## 0.56.29
 
 ### Patch Changes